Refactor to be able to use appstructure in actionhandler
